keep only the composable registering uow
