+++ eva-uow/src/test/kotlin/com/razz/eva/uow/composable/RegisteringUnitOfWorkSpec.kt
@@ -5,22 +5,26 @@ import com.razz.eva.domain.Tag
 import com.razz.eva.domain.TestModel.ActiveTestModel
 import com.razz.eva.domain.TestModel.CreatedTestModel
 import com.razz.eva.domain.TestModel.Factory.createdTestModel
+import com.razz.eva.domain.TestModel.Factory.existingCreatedTestModel
 import com.razz.eva.domain.TestModelEvent
 import com.razz.eva.domain.TestModelEvent.TestModelCreated
 import com.razz.eva.domain.TestModelEvent.TestModelStatusChanged
 import com.razz.eva.domain.TestModelId
+import com.razz.eva.domain.TestModelId.Companion.randomTestModelId
 import com.razz.eva.domain.TestModelStatus.ACTIVE
 import com.razz.eva.domain.TestModelStatus.CREATED
+import com.razz.eva.domain.Version.Companion.V1
 import com.razz.eva.uow.AddEntity
 import com.razz.eva.uow.AddModel
 import com.razz.eva.uow.Clocks.fixedUTC
 import com.razz.eva.uow.Clocks.millisUTC
 import com.razz.eva.uow.DeleteEntity
 import com.razz.eva.uow.ExecutionContext
+import com.razz.eva.uow.NoopModel
 import com.razz.eva.uow.PersistedLookup
 import com.razz.eva.uow.TestPrincipal
 import com.razz.eva.uow.UpdateEntity
-import com.razz.eva.uow.resultOnly
+import com.razz.eva.uow.UpdateModel
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.nulls.shouldNotBeNull
@@ -36,6 +40,63 @@ class RegisteringUnitOfWorkSpec : FunSpec({
     val now = millisUTC().instant()
     val clock = fixedUTC(now)
     val executionContext = ExecutionContext(clock, OpenTelemetry.noop())
+
+    test("Block ending on a registered add produces the same changes a plain UnitOfWork would") {
+        val model0 = createdTestModel("MLG", 420)
+
+        val uow = object : DummyRegisteringUow<CreatedTestModel>(executionContext) {
+            override suspend fun tryPerform(principal: TestPrincipal, params: Params) = changes {
+                add(model0)
+            }
+        }
+        val changes = uow.tryPerform(TestPrincipal, DummyRegisteringUow.Params)
+
+        changes.modelChangesToPersist shouldBe listOf(AddModel(model0, listOf(TestModelCreated(model0.id()))))
+        changes.result shouldBe model0
+    }
+
+    test("with pairs two registrations") {
+        val model0 = createdTestModel("MLG", 420)
+        val model1 = existingCreatedTestModel(randomTestModelId(), "noscope", 360, V1).activate()
+
+        val uow = object : DummyRegisteringUow<Pair<CreatedTestModel, ActiveTestModel>>(executionContext) {
+            override suspend fun tryPerform(principal: TestPrincipal, params: Params) = changes {
+                add(model0) with update(model1)
+            }
+        }
+        val changes = uow.tryPerform(TestPrincipal, DummyRegisteringUow.Params)
+
+        changes.modelChangesToPersist shouldBe listOf(
+            AddModel(model0, listOf(TestModelCreated(model0.id()))),
+            UpdateModel(model1, listOf(TestModelStatusChanged(model1.id(), CREATED, ACTIVE))),
+        )
+        changes.result shouldBe (model0 to model1)
+    }
+
+    test("notChanged registers and returns the persisted model") {
+        val model0 = existingCreatedTestModel(randomTestModelId(), "noscope", 360, V1)
+
+        val uow = object : DummyRegisteringUow<CreatedTestModel>(executionContext) {
+            override suspend fun tryPerform(principal: TestPrincipal, params: Params) = changes {
+                notChanged(model0)
+            }
+        }
+        val changes = uow.tryPerform(TestPrincipal, DummyRegisteringUow.Params)
+
+        changes.modelChangesToPersist shouldBe listOf(NoopModel(model0))
+        changes.result shouldBe model0
+    }
+
+    test("noChanges(result) from the base class is still available") {
+        val uow = object : DummyRegisteringUow<String>(executionContext) {
+            override suspend fun tryPerform(principal: TestPrincipal, params: Params) =
+                noChanges("NOTHING TO DO")
+        }
+        val changes = uow.tryPerform(TestPrincipal, DummyRegisteringUow.Params)
+
+        changes.modelChangesToPersist shouldBe listOf()
+        changes.result shouldBe "NOTHING TO DO"
+    }
 
     test("Entity changes pass through unwrapped while model registrations return Registered") {
         val departmentId = randomDepartmentId()
